Group drivers in dedicated parent module
I think the situation was already a bit unclear, and with more drivers
and more non-driver modules about to be added, it's not going to get
better. Grouping all driver modules into a common parent module makes
things a bit clearer.
