+++ templates/driver/src/lib.rs.tmpl
@@ -12,4 +12,4 @@
 #![no_std]
 #![deny(missing_docs)]
 
-pub use step_dir::\{{ name }::*, *};
+pub use step_dir::\{drivers::{ name }::*, *};
