+++ test-stand/tests/tests/stspin220.rs
@@ -17,7 +17,7 @@ use test_stand::{
         },
     },
     rotary_encoder_hal::Rotary,
-    step_dir::{ stspin220::STSPIN220, StepMode256},
+    step_dir::{drivers::stspin220::STSPIN220, StepMode256},
     test_step,
 };
